Move things around to fix CI error
