+++ python/pyarrow/tests/test_udf.py
@@ -19,12 +19,10 @@
 import pytest
 
 import pyarrow as pa
-import pyarrow._substrait
 from pyarrow import compute as pc
 
 # UDFs are all tested with a dataset scan
 pytestmark = pytest.mark.dataset
-
 
 try:
     import pyarrow.dataset as ds
@@ -39,25 +37,6 @@ def mock_scalar_udf_context(batch_length=10):
 
 class MyError(RuntimeError):
     pass
-
-
-@pytest.fixture(scope="session")
-def unary_func_fixture():
-    """
-    Register a unary scalar function.
-    """
-    def unary_function(ctx, x):
-        return pc.call_function("add", [x, 1],
-                                memory_pool=ctx.memory_pool)
-    func_name = "y=x+1"
-    unary_doc = {"summary": "add function",
-                 "description": "test add function"}
-    pc.register_scalar_function(unary_function,
-                                func_name,
-                                unary_doc,
-                                {"array": pa.int64()},
-                                pa.int64())
-    return unary_function, func_name
 
 
 @pytest.fixture(scope="session")
@@ -614,246 +593,3 @@ def test_udt_datasource1_generator():
 def test_udt_datasource1_exception():
     with pytest.raises(RuntimeError, match='datasource1_exception'):
         _test_datasource1_udt(datasource1_exception)
-
-
-@pytest.mark.parametrize("use_threads", [True, False])
-def test_udf_via_substrait(unary_func_fixture, use_threads):
-    test_table_1 = pa.Table.from_pydict({"x": [1, 2, 3]})
-
-    def table_provider(names, _):
-        if not names:
-            raise Exception("No names provided")
-        elif names[0] == "t1":
-            return test_table_1
-        else:
-            raise Exception("Unrecognized table name")
-
-    substrait_query = b"""
-    {
-  "extensionUris": [
-    {
-      "extensionUriAnchor": 1
-    },
-    {
-      "extensionUriAnchor": 2,
-      "uri": "urn:arrow:substrait_simple_extension_function"
-    }
-  ],
-  "extensions": [
-    {
-      "extensionFunction": {
-        "extensionUriReference": 2,
-        "functionAnchor": 1,
-        "name": "y=x+1"
-      }
-    }
-  ],
-  "relations": [
-    {
-      "root": {
-        "input": {
-          "project": {
-            "common": {
-              "emit": {
-                "outputMapping": [
-                  1,
-                  2,
-                ]
-              }
-            },
-            "input": {
-              "read": {
-                "baseSchema": {
-                  "names": [
-                    "t",
-                  ],
-                  "struct": {
-                    "types": [
-                      {
-                        "i64": {
-                          "nullability": "NULLABILITY_REQUIRED"
-                        }
-                      },
-                    ],
-                    "nullability": "NULLABILITY_REQUIRED"
-                  }
-                },
-                "namedTable": {
-                  "names": [
-                    "t1"
-                  ]
-                }
-              }
-            },
-            "expressions": [
-              {
-                "selection": {
-                  "directReference": {
-                    "structField": {}
-                  },
-                  "rootReference": {}
-                }
-              },
-              {
-                "scalarFunction": {
-                  "functionReference": 1,
-                  "outputType": {
-                    "i64": {
-                      "nullability": "NULLABILITY_NULLABLE"
-                    }
-                  },
-                  "arguments": [
-                    {
-                      "value": {
-                        "selection": {
-                          "directReference": {
-                            "structField": {}
-                          },
-                          "rootReference": {}
-                        }
-                      }
-                    }
-                  ]
-                }
-              }
-            ]
-          }
-        },
-        "names": [
-          "x",
-          "y",
-        ]
-      }
-    }
-  ]
-}
-    """
-
-    buf = pa._substrait._parse_json_plan(substrait_query)
-    reader = pa.substrait.run_query(
-        buf, table_provider=table_provider, use_threads=use_threads)
-    res_tb = reader.read_all()
-
-    function, name = unary_func_fixture
-    expected_tb = test_table_1.add_column(1, 'y', function(
-        mock_scalar_udf_context(10), test_table_1['x']))
-    res_tb = res_tb.rename_columns(['x', 'y'])
-    assert res_tb == expected_tb
-
-
-def test_udf_via_substrait_wrong_udf_name():
-    test_table_1 = pa.Table.from_pydict({"x": [1, 2, 3]})
-
-    def table_provider(names, _):
-        if not names:
-            raise Exception("No names provided")
-        elif names[0] == "t1":
-            return test_table_1
-        else:
-            raise Exception("Unrecognized table name")
-
-    substrait_query = b"""
-    {
-  "extensionUris": [
-    {
-      "extensionUriAnchor": 1
-    },
-    {
-      "extensionUriAnchor": 2,
-      "uri": "urn:arrow:substrait_simple_extension_function"
-    }
-  ],
-  "extensions": [
-    {
-      "extensionFunction": {
-        "extensionUriReference": 2,
-        "functionAnchor": 1,
-        "name": "wrong_udf_name"
-      }
-    }
-  ],
-  "relations": [
-    {
-      "root": {
-        "input": {
-          "project": {
-            "common": {
-              "emit": {
-                "outputMapping": [
-                  1,
-                  2,
-                ]
-              }
-            },
-            "input": {
-              "read": {
-                "baseSchema": {
-                  "names": [
-                    "t",
-                  ],
-                  "struct": {
-                    "types": [
-                      {
-                        "i64": {
-                          "nullability": "NULLABILITY_REQUIRED"
-                        }
-                      },
-                    ],
-                    "nullability": "NULLABILITY_REQUIRED"
-                  }
-                },
-                "namedTable": {
-                  "names": [
-                    "t1"
-                  ]
-                }
-              }
-            },
-            "expressions": [
-              {
-                "selection": {
-                  "directReference": {
-                    "structField": {}
-                  },
-                  "rootReference": {}
-                }
-              },
-              {
-                "scalarFunction": {
-                  "functionReference": 1,
-                  "outputType": {
-                    "i64": {
-                      "nullability": "NULLABILITY_NULLABLE"
-                    }
-                  },
-                  "arguments": [
-                    {
-                      "value": {
-                        "selection": {
-                          "directReference": {
-                            "structField": {}
-                          },
-                          "rootReference": {}
-                        }
-                      }
-                    }
-                  ]
-                }
-              }
-            ]
-          }
-        },
-        "names": [
-          "x",
-          "y",
-        ]
-      }
-    }
-  ]
-}
-    """
-
-    buf = pa._substrait._parse_json_plan(substrait_query)
-    with pytest.raises(pa.ArrowKeyError) as excinfo:
-        pa.substrait.run_query(buf, table_provider=table_provider)
-    assert "No function registered" in str(excinfo.value)
